requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-model-utils==4.4.0
python-dotenv==1.0.1
django-debug-toolbar==4.3.0